feat: Dominio Cidade

+++ algaworks-spring-specialist/src/main/java/com/algaworks/algafood/infrastructure/repository/CidadeRepositoryImpl.java
@@ -4,6 +4,7 @@ import com.algaworks.algafood.domain.model.Cidade;
 import com.algaworks.algafood.domain.repository.CidadeRepository;
 import jakarta.persistence.EntityManager;
 import lombok.AllArgsConstructor;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -34,9 +35,13 @@ public class CidadeRepositoryImpl implements CidadeRepository {
 
         @Override
         @Transactional
-        public void remover(Cidade cidade) {
-            cidade = buscar(cidade.getId());
+        public void remover(Long cidadeId) {
+            Cidade cidade = buscar(cidadeId);
+
+            if (cidade == null) {
+                throw new EmptyResultDataAccessException(1);
+            }
+
             entityManager.remove(cidade);
         }
-
 }
